requirements.txt
numpy==1.19.2
flask
nltk==3.5
faiss-cpu==1.7.0
navec